Remove <T, K, V> generics from traits (#61)
These changes update the runtime traits to elide generic types, which
defined aspects of the command being handled. A policy should be the
only object that is aware of a command's internals. The runtime storage,
therefore, only deals with byte slices in reference to commands.

The `Command` trait is extended to access its serialized bytes. Some
storage traits are also extended in these changes, mostly for accessing
the stored graph and computing the position of objects within the graph.

This also adds an associated type to `Segment` to define an iterator of
commands returned by the added `get_from` method.