Set default-host for self-hosted runners (#80)
Some raspis have 64-bit kernels with 32-bit userspace. This is confusing
for rustup, but rust is smart enough to know the machine needs 32-bit
executables. Rustup, however, bases the host on the kernel info so it
sets the host to be `aarch64-unknown-linux-gnu`. If you attempt to use
any rust components at this point (32-bit executables with `aarch64`
host), they won't run because rustup looks for the 64-bit version, based
on the host.

There are several solutions here: 
1. Set default-host directly on the runners (in this PR) 
2. Directory override for the toolchain to use; but this also requires
overriding the rust version which is currently being set by the
`rust-toolchain.toml`. This solution may not be the easiest to maintain
since we're setting this value in two places, although we are able to
tell where the overrides are coming from.
3. Update the current setup code to be smarter when checking the
environment. The current code just checks the process's architecture,
which is leading to this discrepancy. We have some options for this
solution, or could use a combo of these options:
- There is a `--default-host` option when installing rustup. We could
directly set this to armv7 when running linux.
- `getconf LONG_BIT` tells you the number of bits in the type long int.
This informs us on the size of the OS rather than the kernel (`uname
-m`). We could use this output to verify the host we set when installing
or updating rustup.

(*note*: perhaps this is a long-term goal, but i'm curious if we would
want/need at some point to have a custom build script? I would need to
research the needs for this, I don't know the best solution off the top
of my head. But, we would probably be able to keep it in the repo which
may make it easier to update. Perhaps need a story for this, if we want
it)