all: reduce feature sets and fix warning (#167)
`cargo-all-features` will test both `a,b` and `a` even when `a` enables
`b`, so I added those to the `skip_feature_sets`. This cuts crypto down
from 32 to 14 combinations. We should also always enable `test_util` for
`test-all-features` but not `check-all-features`, but the config for
`cargo-all-features` doesn't allow that currently.