Policy VM fact queries (#127)
The implements the missing pieces required to properly manipulate facts
via the `MachineIO` interface in the VM.

The main change is the addition of a `QueryIterator` associated type in
`MachineIO`. This is a specific type of `Iterator` returned by `query()`
that returns all facts that match the name and partial key. Implementors
of `MachineIO` should also implement `QueryIterator` to provide query
operation.

Other changes:
- I've added some more details to `MachineError`, so most errors will
show you which instruction caused them. This is not generally useful
since most people aren't writing raw VM instructions, but it's a start.
- Most of the (String, *) pairs have been turned into specific tuple
structs for better type checking and ergonomics.
- Reorganized `MachineIO` and `MachineError` into their own modules.
- A few more instructions have been implemented to get this working. In
particular `Dup(n)` was useful for doing fact updates.
- Fixed a couple of compiler bugs.

---------

Co-authored-by: elagergren-spideroak <36516532+elagergren-spideroak@users.noreply.github.com>