Implement Stack trait (#197)
I went around and around on the generics before I settled on this
design. The implementor only needs to implement `push_value()`,
`pop_value()`, and `peek_value()`, and the generic `push()`, `pop()`,
and `peek()` are supplied by the trait. For the `peek()` references, I
created a `TryAsMut` trait (borrowing from [this
crate](https://github.com/Nearoo/try_as/). So you should be able to work
with values on the stack without having to do any intermediate work
converting them to/from the `Value` type.