crypto: updates for APQ (#192)
This PR does a few things:

* Adds postcard `MaxSize` support to anything that implements
`Serialize`/`Deserialize`.

This could help us avoid allocations, although we don't use it yet.

* Makes `Topic` the hash of a string instead of a `u32`.

Mapping topics, which most queues implement as strings, to `u32` isn't
very user friendly. Somebody somewhere has to maintain that mapping. We
tolerate mappings like that in APS (which maps user IDs to `u32`)
because it has stringent performance requirements, but APQ is allowed to
be slower.

Using strings would be ideal. But (1) `String` allocates, which is
something we want to get away from, and (2) `&str` obviously can't be
used as a map key. We could use `&'static str`, but I'm not sure I want
to force users to (effectively) know all possible topics at compile
time.

Hashing `&str` into `[u8; 16]` lets users use strings and lets us use
fixed-size objects. I chose SipHash128 (thanks, @ycarmy) because it's a
fast cryptographic hash (*). We cannot risk collisions, so we need a
hash with a negligible risk of collisions. We don't use `E::Hash`
because this is _not for cryptographic purposes_. We simply value the
collision resistance provided by a cryptographic hash.

*: well, it's a MAC, and it's not actually designed to be collision
resistant, but the 128-bit output space gives us a collision risk of
2^-64, which is good enough.

I'm sure there might be other possibilities. We could probably make
`Topic` a generic type that implements `AsRef<[u8]> + Sized` or
something. I'm open to ideas, since I'm not super thrilled about this
implementation.

If we stick with the hash we may also want to consider reducing the size
to 16 bytes. It'll let us make use of some optimizations (e.g., ZeroMQ
won't need to allocate memory for it since it's under 30 bytes) and
using less network resources is always good thing.

* Removes the record name from `SenderSigningKey::sign`.

It forces receivers to decode the record before checking the signature,
which isn't really a good idea. (It opens the door to DoS attacks, etc.)
The encoding format has to know the record type anyway, so that
information is already part of the signature.

* Do not use `ExportedData` for encoding/decoding `Signature`.

Our current method that `ExportedData` strongly prevents cross-protocol
attacks, but it also makes it very difficult to use `Signature` with
`no_alloc` (we're essentially concatenating three fixed-size fields with
one variable size field).

We don't use `ExportedData` for `ExportedGroupKey` and
`ExportedTopicKey` because they use the same information that
`ExportedData` provides (engine ID, suite IDs, etc.) as info/AD to HPKE.

We don't use `ExportedData` for `Encap` for similar reasons. (However,
it's ciphertext, so there might be key committing issues, so I'll need
to rethink this one.)

I think we can ignore cross-protocol attacks for `Signature`. We require
EUF-CMA security for signatures, and if an attacker can present a
cross-protocol forgery then the signature algorithm clearly isn't
EUF-CMA secure. I suppose it's possible for the underlying crypto
library (BoringSSL, BearSSL, a Type 1 encryptor, etc.) to get confused
by an invalid signature (because of a bug or whatever), but I'm not
really sure we should try and protect against that case.

Fixes #222

---------

Signed-off-by: Eric Lagergren <elagergren@spideroak-inc.com>