crypto: rewrite tests using macros (#226)
Ideally, we'd use something like Go's sub-tests. But we don't have that
in Rust, so use macros instead. It expands to about ~350 tests per
backend, so ~1,000 with `rust`, `bearssl`, and `boringssl`. This also
has the benefit of letting all the tests run concurrently (bounded by
whatever number of threads libtest decides to use, of course).

While we're at it, make `cargo-all-features` skip some unnecessary
features.

CI now finishes in about 3.5 hours on the slow arm and arm64 runners, so
a ~50% improvement.

Fixes #98
Updates #160
Updates #225

---------

Signed-off-by: Eric Lagergren <elagergren@spideroak-inc.com>