Add clippy check without `--tests` (#276)
Cargo will build only one version of a crate if it can. Crates are
compiled with "default" enabled unless using `--no-default-features`. In
the case a crate is a dependency of multiple crates within the
workspace, cargo will compile the dependency crate with unified features
(as determined by the dependent crates).

For example, say we have a workspace with crates A, B and C. If crate A
is a dependency of crate B with `features = "foo"` and crate A is a
dependency of crate C with `features = "bar"`, cargo will compile crate
A with both of these features enabled when building the entire project
(from workspace root). Further, crate A will have "default" enabled
unless both crate B and C set `default-features = false` for the crate A
dep in their respective Cargo.tomls (or if crate A does not specify
"default" features in its own Cargo.toml).
When building crate A on its own (either from inside crate A or using
`-p A` from workspace root), "foo" and "bar" are not compiled in if they
are not "default" (and if they are, they can be disabled using the
`--no-default-features` flag).

See feature-unification docs here:
https://doc.rust-lang.org/cargo/reference/features.html#feature-unification
and more info on how cargo resolves/decides features here:
https://doc.rust-lang.org/cargo/reference/resolver.html#features

Solution: Validate each individual package separately. This option would allow
for checking more fine-grained feature sets in specific crates and
removes the feature unification stuff when compiling from the workspace
root. Separately checking would still enable "default" unless
`--no-default-features` is used (or the individual package does not
define "default" features). _This may be more favorable as it does not
lead to crates being compiled/checked with features pulled in by other
crates that depend on it._ Also, on the individual level, default is
easier to turn off and doesn't require setting `default-features =
false` in crates that depend on it.