policy-lang: split into multiple crates (#298)
There are two new crates: `policy-ast` and `policy-vm`.

`policy-vm` contains the virtual machine.

`policy-ast` only contains the AST. It's its own crate because it's
shared between `policy-lang` and `policy-vm`. I think we can refactor
this to get rid of the crate, but it wasn't worth doing in this PR.

Both `policy-ast` and `policy-vm` only require alloc. We can probably
get rid of this requirement in the future.

I added some simple documentation.

Fixes #297

---------

Signed-off-by: Eric Lagergren <elagergren@spideroak.com>