policy-lang: Better Error Reporting (#175)
This improves errors reported from the VM to include location
information and text from the original source. The original source is
kept inside a CodeMap object, which is optional inside the Machine. It
maps instructions to text ranges and is used when an error is generated.

To propagate location information through the parser and compiler, a new
generic `AstNode<T>` type was added to contain this information. It
wraps many of the AST objects. In the compilation stage, this
information is mapped to an instruction range inside the CodeMap object.

I also expanded and clarified the machine error types and wrote some
test cases to check that they are produced when they should be.

Closes #149.