runtime: split off fact index from segment (#342)
Fixes some problems with facts in the storage and client. Impact to
other code is fairly low, most of the diff is in memstorage.

An important change to be aware of is that `Engine.call_rule` no longer
adds commands to the perspective, so the calling code (e.g. client.rs
and anything directly testing engines) must do so.

`FactIndex`es are split off from `Segments`, and `Perspectives` are
split into `trait FactPerspective` and `trait Perspective:
FactPerspective`. This is because braids do not need to store the actual
commands and because branching from the middle of a segment needs to
create a new prior fact index, to support branching from the middle of
that branch..

Closes #281.