ci: improve deny config (#397)
New versions were causing error in CI so I took a closer look at the
config.

Some of our conflicts were coming from OS specific dependencies like
windows and redox, so setting the target list means we don't have to
worry about those.

`rtrack` doesn't need to be considered for duplicate dependencies since
it would only be used as a dev-dependency. (If/when we move it to
another repo and use it as a dev dependency `cargo-deny` will ignore it
for this anyway)

The two above changes reduces the number of duplicates we need to
ignore. I also set the version, as recommended.

I removed an unused license too.