Runtime/Policy VM Integration (#190)
Keep in mind that this is still very much a minimal prototype
implementation of `engine::Policy` - it doesn't do anything with
`Payload` or `serial()`, doesn't implement FFI, and is probably doing
things slightly wrong besides. I fully expect that this will be heavily
modified or completely hacked up and moved into a different module. It
only works with the `TestEngine` and `TestSink` in the tests.

I added an escape hatch in `ClientState` to provide crate access to the
`provider`, for verifying in the fact modification by poking at the
storage directly.

From the prior version of this PR:

> The main limitation of this implementation is that you can't yet do
partial key queries since there is no partial key query interface in
Perspective. Error reporting is also... bad. Any policy that executes a
panic instruction (like a failed check or unwrapping a None result from
a query) will simply be reported as NotAuthorized.