Add separate error crate (#404)
Adds a `trouble` crate that re-exports or defines `Error` as necessary,
so the other crates don't need to use `cfg` to conditionally `impl
Error`. This has a `std` feature and will detect nightly in a
buildscript to enable `error_in_core` so that each depending crate
doesn't need to have an `error_in_core` feature.