runtime: add lifetime to `Policy::Actions` (#410)
Instead of just changing the vm policy action to a slice, I'm changing
it to a `Cow<'a, [Value]>`. This means I can choose to pass in an owned
vec in the interface codegen when the arguments could be expensive to
clone. (This is actually a bit worse if they are all empty, since it
should require no allocation, but I _could_ fix this by checking if all
are empty...)

```rust
fn expensive(bytes: Vec<u8>) {
    client.action("expensive", Cow::Owned(vec![Value::from(bytes)]));
}

fn cheap(int: i64) {
    client.action("cheap", Cow::Borrowed(&[Value::from(int)]));
}
```

I changed the vm's action methods to take an iterator to support this.