runtime: fix sync test flake (#453)
The flake error was "addr in use". I changed the tests_api to use
unassigned ports.

I also improved some of the error handling and am now logging the quic
sync errors. I don't think that's necessarily the only way we should
report those errors but it's an improvement for now.

Closes #442.