policy-vm: integrate seal and open (#514)
This integrates the seal and open blocks into `VmPolicy`, so that they
are called at the appropriate times. This integration is not yet
complete, but it is kind of working so I thought it was time to put
eyeballs on it.

I have rearranged the guts of the VM to make this work. In summary:
- `CommandContext` is now an enum with variants for each context in
which policy code can run (action, policy, recall, seal, open). It
probably needs to be renamed and is certainly not complete.
- `CommandContext` is now passed into the VM as a shared ref when
creating `RunState`. This means each `RunState` should only exist for
one call into the VM. I want to remove the need for VM users to manage
`RunState` at all, but this is ongoing.
- `&mut E` (`where E: Engine`) is now divorced from `CommandContext`. It
is now owned and managed by `MachineIO` implementations so the VM
doesn't have to manage it.
- Likewise, the FfiModule trait's `call()` now accepts parameters for
both `&CommandContext` and `&mut E`

Closes #331, closes #428