runtime: large graph test (#524)
I left in a test with 100 commands so it doesn't take too long to run. I
manually tested up to 1000 commands with a variety of clients, command
chances, and sync chances. With more than 1000 commands it starts to
cause stack overflows.

All of the generated rules are printed so if there's an error it should
be possible to reproduce. At some point we'll probably want genuinely
large graphs with 100s of thousands of commands, but this test should be
a reasonable start for our current work.

Fixes #492

---------

Co-authored-by: Jonathan Dygert <jdygert@spideroak.com>