Move canaries to workspace and fix canary-alloc (#630)
Canaries were outside of the workspace because most commands should not
interact with them, but excluding them from `default-members` pretty
much solves this. Having the canaries in the workspace is much less
awkward regarding `target` and `Cargo.lock`.

I also had to add back the `extern crate`s to `canary-alloc`, for it to
actually complain about missing global allocator when using alloc
dependencies.