Changed `compile_from_policy` to `Compiler` builder (#696)
This changes the `compile_from_policy` to the `Compiler` builder and can
be thought of as running the regexp
`compile_from_policy\((&\w+),\s*(&?\[.*?\]|(?:\w+::)*\w+)\)` and
replacing it with `Compiler::new($1).ffi_modules($2).compile()`. There
is no change in functionality. This also removes `compile_from_policy`
as it's redundant