760/Add configurable FFIs to model (#787)
This PR handles adding configurable FFIs to the model. We want to have
the ability to use the full power of foreign functions in the runtime
model, while still allowing our selves the convenience of having a
simple interface when that power is not needed. These changes give us
the ability to pass in the minimum configuration needed to create
clients.