runtime: add `testing` module (#827)
The no-std `testing` module exposes (some) of the runtime tests as a
library so that it can be used externally. The primary impetus is #815
where the new `vxtest` crate (which is no-std) needs to perform the
runtime tests.

The DSL tests now use JSON instead of YAML. While YAML is a little nicer
to use, there isn't a no-std YAML crate. So, use `serde_json` which
supports no-std.

This PR also cleans up the `crypto`'s CSPRNG so that it can be used by
`runtime`'s DSL tests instead of `rand`.

For #815

---------

Signed-off-by: Eric Lagergren <elagergren@spideroak.com>