policy-compiler: Check Types (#879)
The core of this is `calculate_expression_type()` in the new
`compile::types` module. It recurses down an expression tree, attempting
to calculate the resulting type. The type is expressed as a `Typeish`
enum, which holds either a concrete type, or `Indeterminate`, which
means the checker couldn't determine the type. It also checks that the
types of sub-expressions are correct in their context, and will error
with a `TypeError` if they are not.

The calculated type is now additionally returned from
`compile_expression()`. In things like `let` statements, the type is
associated along with the identifier in an `IdentifierTypeStack`, which
tracks identifiers across scopes and lets the type checker look up those
identifiers' types when they occur in expressions. Expression types are
additionally used in `check`, `return`, and `emit` to check that they
are given the correct type.

An `Indeterminate` result matches any potential type with `is_maybe()`,
which means type checking is not yet strict. A small number of things
cannot yet be determined - the `None` literal, and the `deserialize()`
builtin. Deserialize could be determined, it's just that the information
about the current command isn't available inside the type checker.
`None` is a puzzle, and I think it might wind up being a more narrowly
defined kind of `Indeterminate`.

In order to support type checking FFI, the derive bits were modified to
include the structs in the `SCHEMA` member, which are then defined in
the machine for compile- and run-time checking. Since we're now checking
the `Envelope` type in `seal` and `open`, there is a hack that defines
the `Envelope` struct when the envelope FFI is not used (basically only
in tests).

Naturally, this caught some fast and loose behavior in tests and FFI
definitions.

Closes #772
(also completes most of #773)
Fixes #884