runtime, policy-vm: Report command IDs and recall status in effects (#921)
This adds a command ID and a recall flag to the `effect()` method of the
`Sink` trait. Likewise, fields were added to `VmEffect`. This
necessitated a related type which is identical to `VmEffect` without
these extra fields - `VmEffectData`. `PartialEq` is implemented between
these so that existing `assert_eq!(foo, vm_effect!(...))` tests will
continue to work as before.

This also patched up some things that were missing because Recall hadn't
effectively been tested:
- Proper CommandContext wasn't set on recall execution, necessitating
some hackery to change the context in the middle of policy evaluation.
I'm not very pleased with it, and am open to suggestions.
- ~~There was no way to set priorities, so it was impossible to create a
reliable scenario where recall would happen. I added another hack to set
a priority map. I expect this will eventually be hooked into the
attributes system to populate priorities automatically. For now it is
only available with the `testing` feature.~~ Attributes have been
merged, and this is now using them.

Closes #839