Add Compiler Tracer (#955)
This adds a `TraceAnalyzer` object which is generic over a `Tracer`
implementation. It brute-force traverses the compiled instruction paths
to prove facts about the code. While doing so it records the instruction
path so that errors can be traced back to the specific series of
statements that led to it (associating this path with statements is
currently unimplemented).

I have also implemented a compiler flag `stub_ffi`, which doesn't
compile or check FFI invocations. This allows you to use the standard
policy compiler/VM explorer to prove things about policies that use
custom FFIs.