crypto: update to `spideroak-crypto` v0.6 (#300)
Change some usages of `tuple_hash` to `zerocopy` structs. Whether this
is better depends on the KDF. For a KDF like HKDF, this is generally
better if the output key size is less than or equal to the digest size.
For a KDF like [KMAC], this is almost always better. (NB: It also
assumes that all struct fields are fixed size and unambiguous.)

This PR only changes calls to `tuple_hash` that were passed to HPKE.
Other usages will be changed in later PRs.

[KMAC]:
https://nvlpubs.nist.gov/nistpubs/SpecialPublications/NIST.SP.800-108r1-upd1.pdf

For #298

---------

Signed-off-by: Eric Lagergren <eric@ericlagergren.com>
Signed-off-by: Eric Lagergren <elagergren@spideroak.com>
Co-authored-by: Eric Lagergren <eric@ericlagergren.com>