policy-compiler: remove `Typeish::Probably` and use `Never` type (#347)
The only remaining uses of indeterminate were both "panics" so it should
really be a "never" type, which can unify with anything but will never
be actually instantiated. I also added a panic instruction when ffi is
stubbed instead of just doing nothing which would cause weird things to
happen.

---------

Co-authored-by: Steve <ssweetney@spideroak.com>