runtime: check persistence when calling actions (#380)
Ensures that persistent actions are used on-graph and ephemeral actions
are used off-graph.

I changed the API to use `ActionPlacement` and `CommandPlacement`
instead of `Persistence` and `OnRecall`. I think it's more clear this
way to describe how the methods are being called and have the
implementations determine the correct behavior from that.