Enforce More Clippy Lints (#385)
This adds a number of lints that are allowed by default to our warnings
and fixes all the instances. This should allow our codebase to be a lot
more readable at a glance.

The main changes are enforcing `Self`, better use of .clone(), making
sure we use semicolons consistently, using `as _` for traits that are
only used anonymously, and making sure we use core::ptr instead of raw
pointer casts.

---------

Co-authored-by: Jonathan Dygert <jdygert@spideroak.com>
Co-authored-by: Geoff <gknopf@spideroak.com>