policy-compiler: split type checking and instruction generation (#500)
This PR adds a typed intermediate representation. Compiling any body
(function body, policy block, etc), is now split into two steps. We
first typecheck and generate the IR, then consume the IR and generate
instructions.

Currently, the IR is fairly similar to the ast. It duplicates all
expressions and statements, adding a type to the expressions. There are
a few other changes which help move some checks up into the typechecking
step. The IR does not need to exactly match the ast and can evolve
independently.